Set up initial sample program that contains the most barebones setup

+++ ToucheTools/Models/SpriteImageDataModel.cs
@@ -2,10 +2,10 @@
 
 public class SpriteImageDataModel
 {
-    public int Width { get; set; }
-    public int Height { get; set; }
-    public int SpriteWidth { get; set; }
-    public int SpriteHeight { get; set; }
+    public short Width { get; set; }
+    public short Height { get; set; }
+    public short SpriteWidth { get; set; }
+    public short SpriteHeight { get; set; }
     public byte[,] RawData { get; set; } = new byte[0, 0];
     public byte[,] DecodedData { get; set; } = new byte[0, 0];
 }
